Refactors file locations and updates a few styles
