+++ lib/responseDownloadFormats/html-aggregated/index.ts
@@ -1,6 +1,6 @@
 import { renderToStaticMarkup } from "react-dom/server";
 import { FormResponseSubmissions } from "../types";
-import { ResponseHtmlAggregated } from "../html/components/ResponseHTMLAggregated";
+import { ResponseHtmlAggregated } from "./components/ResponseHTMLAggregated";
 
 // TODO add language param to API call probably as a query or from header info
 export const transform = (formResponseSubmissions: FormResponseSubmissions, lang = "en") => {
