+++ lib/responseDownloadFormats/html/tailwind.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  content: ["./**/*.{js,ts,jsx,tsx}"],
+  content: ["./**/*.{js,ts,jsx,tsx}", "../html-aggregated/**/*.{js,ts,jsx,tsx}"],
   theme: {
     extend: {
       fontFamily: {
